refactor(analysis): move adapters into FSD api + lib segments
REST gateway & service stub -> api/ (HTTP-facing); SystemClock and the pure
toAreaSelection / analysisResultToWidgets mappers -> lib/. Imports rewired to
reach ports/types in model/.
