+++ src/features/analysis/lib/__tests__/to-area-selection.test.ts
